integrate eloquent migrations

+++ app/migrations/2014_11_04_133750_create_activity_tags_table.php
@@ -1,35 +1,42 @@
 <?php
 
+use Illuminate\Database\Capsule\Manager as Capsule;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateActivityTagsTable extends Migration {
+class CreateActivityTagsTable extends Migration
+{
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-        Schema::create('activity_tags', function(Blueprint $table) {
-            $table->integer('activity_id')->unsigned();
-            $table->integer('tag_id')->unsigned();
+    protected $table = 'activity_tags';
 
-            $table->foreign('activity_id')->references('id')->on('activities')->onDelete('cascade');
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');;
-        });
-	}
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        if (!Capsule::schema()->hasTable($this->table)) {
+            Capsule::schema()->create($this->table, function(Blueprint $table) {
+                $table->integer('activity_id')->unsigned();
+                $table->integer('tag_id')->unsigned();
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-        Schema::table('activity_tags', function(Blueprint $table) {
-            $table->drop();
-        });
-	}
+                $table->foreign('activity_id')->references('id')->on('activities')->onDelete('cascade');
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
+            });
+        } else {
+            
+        }
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Capsule::schema()->dropIfExists($this->table);
+    }
+
 }
