+++ app/migrations/2014_11_04_133802_create_timeslice_tags_table.php
@@ -1,35 +1,42 @@
 <?php
 
+use Illuminate\Database\Capsule\Manager as Capsule;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateTimesliceTagsTable extends Migration {
+class CreateTimesliceTagsTable extends Migration
+{
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
-	public function up()
-	{
-        Schema::create('timeslice_tags', function(Blueprint $table) {
-            $table->integer('timeslice_id')->unsigned();
-            $table->integer('tag_id')->unsigned();
+    protected $table = 'timeslice_tags';
 
-            $table->foreign('timeslice_id')->references('id')->on('timeslices')->onDelete('cascade');
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
-        });
-	}
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        if (!Capsule::schema()->hasTable($this->table)) {
+            Capsule::schema()->create($this->table, function(Blueprint $table) {
+                $table->integer('timeslice_id')->unsigned();
+                $table->integer('tag_id')->unsigned();
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
-	public function down()
-	{
-        Schema::table('timeslice_tags', function(Blueprint $table) {
-            $table->drop();
-        });
-	}
+                $table->foreign('timeslice_id')->references('id')->on('timeslices')->onDelete('cascade');
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
+            });
+        } else {
+
+        }
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Capsule::schema()->dropIfExists($this->table);
+    }
+
 }
